Beam analysis tool.
Version 1 of the beam analysis tool. Takes an input OBSID.metafits file, outputs the 1st to 4th order log-polynomial residuals of the beam across the whole sky. Optionally outputs a table of the high residual pixels in the beam map, these occur in regions of local minima and maxima, which change location as a function of frequency.